+++ src/Logger.php
@@ -76,9 +76,9 @@ final class Logger
     ];
 
     /**
-     * @var Monolog|null The monolog instance
+     * @var Monolog The monolog instance
      */
-    protected $instance = null;
+    protected $instance;
 
     /**
      * Create a Monolog instance and attach a handler
@@ -108,13 +108,11 @@ final class Logger
     /**
      * Send log message to output interface
      *
-     * @param string $message Message to output
-     * @param mixed  $context Some context around the log
-     * @param int    $logType The log type
-     *
-     * @return mixed
+     * @param string   $message Message to output
+     * @param mixed    $context Some context around the log
+     * @param int|null $logType The log type
      */
-    public function log(string $message, $context = null, int $logType = Monolog::INFO)
+    public function log(string $message, $context = null, ?int $logType = null)
     {
         if (is_int($context) and is_null($logType)) {
             $logType = $context;
@@ -123,6 +121,10 @@ final class Logger
 
         if (!is_array($context)) {
             $context = is_null($context) ? [] : [$context];
+        }
+
+        if (!in_array($logType, $this->instance->getLevels())) {
+            $logType = Monolog::INFO;
         }
 
         return call_user_func([$this->instance, $this->logTypes[$logType]], $message, $context);