+++ src/Queue.php
@@ -17,7 +17,7 @@ namespace Resque;
  * @package Resque
  * @author Michael Haynes <mike@mjphaynes.com>
  */
-final class Queue
+class Queue
 {
     /**
      * @var Redis The Redis instance
@@ -79,7 +79,7 @@ final class Queue
      *
      * @return Job job instance
      */
-    public function push($job, ?array $data = null, ?string $queue = null): Job
+    public function push($job, ?array $data = [], ?string $queue = null): Job
     {
         if (false !== ($delay = Config::read('default.jobs.delay', false))) {
             return $this->later($delay, $job, $data, $queue);